add sm21 sm22 sm23

+++ sm13/sm13-4.c
@@ -1,13 +1,17 @@
 #include <stdio.h>
+
+
 int main() {
 
-    unsigned long long sz = 0;
+    size_t sz = 0;
     unsigned int left, right;
 
     while (scanf("%x-%x %*[^\n]%*c", &left, &right) != EOF) {
     	sz += right;
     	sz -= left;
     }
-    printf("%llu", sz);
+    printf("%lu", sz);
     return 0;
 }
+
+// В задаче не нужны 64-битные типы
